moves triangles into navbar component
